Achieve true 3-file architecture
Simplified the codebase to exactly 3 core operational files as intended:

1. **github_client.py** - GitHub API client with authentication
2. **workflow_cli.py** - CLI and all workflow management logic (merged from workflow_manager.py)
3. **app.py** - Webhook server

Changes:
- Merged WorkflowManager class into workflow_cli.py to eliminate the extra file
- Moved setup script to scripts/setup.py (it's a one-time utility, not core)
- Removed SPEC.md (original requirements, not needed for operation)
- Updated README to reflect the true 3-file architecture

The system now has exactly 3 core Python files for all operations, with supporting configuration and setup files clearly separated.
